Move the text segment out to rich_text_objects
